Add basic error handling and some fixes

+++ services/groupService.js
@@ -8,7 +8,10 @@ const groupService = (groupModel, lessonModel, slotModel, dayModel) => {
     }
 
     const editOne = async (groupName, newValues) => {
-        const group = await groupModel.findOne({ groupName: groupName }).exec();
+        const group = await groupModel.findOne({ name: groupName }).exec();
+
+        if(!group)
+            throw new Error('Group not found');
 
         for(let key of Object.keys(newValues)) {
             group[key] = newValues[key];
@@ -18,13 +21,17 @@ const groupService = (groupModel, lessonModel, slotModel, dayModel) => {
     }
     
     const deleteOne = async (groupName) => {
-        const group = await groupModel.findOne({ groupName: groupName }).exec();
+        const group = await groupModel.findOne({ name: groupName }).exec();
+
+        if(!group)
+            throw new Error('Group not found');
+
         await group.deleteOne();
     }
 
     const getSchedule = async (groupName, date = undefined, slot = undefined, time = undefined) => {
         let filter = {};
-        filter.group = groupName;
+        filter.groups = groupName;
 
         if(date) {
             filter.day = date.getDay();
@@ -56,7 +63,7 @@ const groupService = (groupModel, lessonModel, slotModel, dayModel) => {
 
     const getRemainingSchedule = async (groupName) => {
         let filter = {};
-        filter.group = groupName;
+        filter.groups = groupName;
 
         let date = new Date();
         let time = date.getHours() < 10 ? `0${date.getHours()}` : `${date.getHours()}`;
